--- src/main/java/racingcar/ui/RacingGameParameterInputView.java
@@ -1,34 +1,0 @@
-package racingcar.ui;
-
-import racingcar.vo.RacingGameParameter;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Scanner;
-import java.util.stream.Collectors;
-
-public class RacingGameParameterInputView {
-    private RacingGameParameterInputView() {
-    }
-
-    public static RacingGameParameter readRacingGameParameter() {
-        Scanner scanner = new Scanner(System.in);
-
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        String carNamesString = scanner.nextLine();
-        System.out.println("시도할 회수는 몇 회 인가요?");
-        int tryCount = scanner.nextInt();
-        System.out.println();
-
-        scanner.close();
-
-        List<String> carNames = convertCarNamesStringToList(carNamesString);
-        return new RacingGameParameter(carNames, tryCount);
-    }
-
-    private static List<String> convertCarNamesStringToList(String carNamesString) {
-        return Arrays.stream(carNamesString.split(","))
-                .map(String::trim)
-                .collect(Collectors.toList());
-    }
-}
