+++ src/main/java/racingcar/view/RacingGameOutputView.java
@@ -1,6 +1,6 @@
-package racingcar.ui;
+package racingcar.view;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 
 import java.util.List;
 import java.util.stream.Collectors;
